Packaging: subsystem: don't remove chroot and conf file during upgrade process
git-svn-id: file:///home/code/svn/ovd/trunk@5488 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ packaging/rpm/ovd-subsystem/opensuse-11.1.spec
@@ -43,11 +43,13 @@ cp init/suse/ulteo-ovd-subsystem $INITDIR
 service ulteo-ovd-subsystem stop
 
 %postun -n ulteo-ovd-subsystem
-SUBCONF=/etc/ulteo/subsystem.conf
-CHROOTDIR=/opt/ulteo
-[ -f $SUBCONF ] && . $SUBCONF
-rm -rf $CHROOTDIR
-rm -f $SUBCONF
+if [ "$1" = "0" ]; then
+    SUBCONF=/etc/ulteo/subsystem.conf
+    CHROOTDIR=/opt/ulteo
+    [ -f $SUBCONF ] && . $SUBCONF
+    rm -rf $CHROOTDIR
+    rm -f $SUBCONF
+fi
 
 %clean -n ulteo-ovd-subsystem
 rm -rf %buildroot
